Fixing wrong code with the lowpass algorithm
The calculations of samples_per_antenna_2 and _3 in rx_dsp_chain was wrong. It was calculating the total number of frequency data set samples for each antenna. I fixed this issue.

In the lowpass algorithms, the offset calculations were wrong. The frequency data set offset into the input samples was wrong and needed to account the total number of antennas. The output sample offset was wrong and still included the calculation for the bandpass.

I also cleaned up various comments.