Added data taps for intermediate decimation stages
I added additional buffers to be able to tap the outputs of each stage of decimation. In order to test decimation, I also started the framework for how data_writing will work. I created a protobuf as a proxy to another application. The protobuf will hold the ACFs and meta data, and will optionally hold debug data from each stage that can be written to file.