Improved numerous simulation parameters
Since we are now using ATR, I pad the beginning and end of the pulse with 0s so that ATR begins and ends when the atten window needs to occur. I also improved the calculation for number of recv samples for a given sequence. I clear the protobuf fields that only need to be sent once.