Created simulation of all block communication
We created a simulation of all the blocks talking to each other using a ROUTER/DEALER configuration. For now we just mimic the communication used in the current model. Now that the first step is done we can move on to simulating what kind of communication is needed using brian.