Added phase correction and sample dropping
As part of the modified frerking filting translation, the phase term pulled out of filtering needs to be multiplied by the output samples. This pull request addresses this.

This PR also adds in the calculation to drop samples if they may introduce edge effects from using 0 padded values that are used if the filter extends past the length of the input samples.