Converted the driver over to use continuous rx
In order to use a continuous rx system, the system configures a start time and then starts endlessly receiving into a ring buffer. The tx side calculates the sample it would need when a new pulse sequence starts and then those samples are copied to a shared memory section for processing.