Fixed different amplitudes issue
We noticed that the amplitudes of the signal were much different on different channels. We traced the issue to the fact that the tx channel numbers were being used when tuning the rx channel. What we were seeing on the low amplitude channels was actually just leakage from the tx chain, and those channels were still at DC.