Fixed issue of filtering first pulse
I fixed this issue of the first pulse being filtered. The samples were being filtered because we were not propagating enough samples through the filter in the start or end. We were getting caught in the roll off. I went through the changes needed to move samples from ringbuffer into gpu directly from the dsp program. This lets me use filter information that is already calculated and it removes some of the work needed from the driver application.

The driver, dsp application, protobufs and simulators needed to be changed for this. I also made some helper modifications to brian to use it for router only mode