Making changes to optimize speed
Had to adjust the flow control so that we weren't waiting on the samples
to be copied to device. That adds several ms of unneeded delay. This
involved removing the request from radar control to dsp and making the
sequence timing require a third variable to trigger a new sequence.

Radar control has a bug with 'isarepeat' that needs to be fixed. This
causes a slow down as the pulses are regenerated every sequence. It was
also causing the device to be reconfigured every time which added lates.

Data write was modified to be slightly async as well since the IO could
take longer than a sequence and would cause the system to lock up while
waiting for an ACK.