Correlating tx and rx samps, finding phase offset
Updated test_beamforming.py to find the phase offset between tx and rx
samples for each pulse in the decimated samples data. The phase offset
for each pulse's sample was consistent (error of approx 0.0003 rads,
0.02 degrees max between pulses). At 10.5 MHz this corresponds to 5 ps
error in the ADC so this is good.

Corrected a problem with a floating point error in num_samples in
write_samples_to_file in samples_building.py that resulted in a missed
sample in the sequence.