Moving towards threaded debug mode write samples
Working on getting threading in radar_control to write the samples to
file in debug mode in a separate thread. Tkinter is not thread safe so
removed the plot_samples call in make_samples in sample_building, and
removed all plotting functions from sample building over to a new file
under testing (which is more appropriate).

Added semaphore and threading to radar_control but currently still
having issues with something blocking.