More beamform testing
We calculated the phase offset between tx and rx signals for one box and
then saved those offsets to a file to use as calibration when looking
for the beam phase when using more boxes. When we began to look at more
boxes, we noticed some strange behaviour where the pulse sequences are
offset from the start of where they should be. We tried some runs with 0
phase applied(straight down the boresight) and it looks like it is still
happening. Keith tried some runs looking at individual IQ and it looks
like its a ringbuffer index issue. Keith is going to work on improving
data writing capabilities so we can more easily collect data.