Improving performance
Writing out raw RF samples took unrealistically long for anything more
than 1 antenna in the old format. Lots of parsing and deserializing was
done in sequentially as well. I changed the way that DSP and data write
interact with raw samples. DSP now sacrifices a small bit of time to
copy the RF samples to a shared memory section that data write can
optionally write directly to file, or can delete if not needed. This
gives the option to more easily work with raw rf in testing or if needed
for research.

In data write, the design was changed to create a function that parses
out the received data for each desired type. The thread then spawns a
separate process to run that function and write out the data. This way
each data product can truly be created concurrently.