Fixing offset bug
We now use timestamps to determine where in the buffer to index next for
samples. In the case of an overflow, UHD does not return the number of
samples lost. We need to use timestamps to determine how far in the
buffer we need to move along to account the missing samples.