Profiling bug updates
Updated the radar_control_input_sim to work in python3 (with send_bytes
function from socket_operations.py). Added a sequence_time field in
radar_control send_metadata function; only sending 0.0 but can be
updated in the future.

Updated the Readme with the skeleton of how to compile the protobufs for
python (I've looked it up too many times!).

Removed a useless file