Added one_pulse_only to AveragingPeriod to fix bug
Implemented a method to speed up basic experiments where the
AveragingPeriod has a member one_pulse_only that is set to True if the
averaging period only has a single unique pulse in it. This allows radar
control to send all repeats after the first pulse in the averaging
period (or integration time). This allowed me to fix the bug in
radar_control where it was only ever sending repeats after the first
pulse sequence ever called by radar_control (an error if another beam
direction is ever sent), but still maintain the speed to get 34 averages
out (using the radar_control_input_sim).