Added slice keys lag_table, rsep, and comment
These keys will be used for acfs and will be supplied to datawrite to be
written in the files.

Changed the default behaviour of slices so that acf is not default
(requires additional fields). Acfint and xcf are default if acf is set
to True, otherwise all correlation is not default. The lag_table will be
calculated if not supplied, as will rsep.