Fixed problem in bfiq datawrite, testing
Scope problem in datawrite caused interferometer data to be written as
both main and interferometer on the first sequence of an integration
period in the bfiq, that has been fixed.

Updated test_beamforming with more plotting functions that are easier to
call, starting to incorporate tx and rawrf data.

Fixed an exception handling syntax error in experiment_prototype