Fixed problems with updating experiment.
Problems fixed while testing update method and edit_slice methods. Also
edited radar_control so that it will not search for an experiment
immediately after it loaded a new experiment.

Edited experiment_handler so that it will build an updated experiment
(change_flag was made into a global).

Edited experiment_prototype so that it clears old scan_objects when it
runs the build_scans method.