First_rx_sample time calculated
Calculated the location of the first rx sample from the start of the tx
samples. This is a sample offset from the start of the samples passed to
the driver, at which the receiver should sample and use as sample 0 in
the output data. This is to align the rx samples with the first pulse.

Still TODO: pass this to the driver.

Also, removed the debug_path for writing debug samples, as that is now
done in data_write and was missed when that changed occurred.