Cleaned up some config related options
I cleaned up some config related items. I added the octoclock addr to
the config and changed the usrp code to use that. I also removed the
unused fields from the config file and code. I noticed that the n200
code was not using the uhd streaming types from the options, so I
changed that hard coded stuff.