Passing information using transmit_metadata dict
Creating a dictionary in experiment_prototype called transmit_metadata
and using that to pass the information down to the sequence level with
information required for building samples. This information comes from
both config file and the experiment instantiation.

Also made rxctrfreq and txctrfreq non-modifiable to avoid any
transmitting errors at this point.