Changes for channels = config always
Implemented changes in radar_control and the driverpacket protubuf (with
some new comments in sample_building and sigprocpacket protobuf) because
the tx_antennas list will no longer be sent to the driver but instead
radar_control sends a number of transmit samples for every antenna as
available from the config file. Anything not wanted to be transmitted on
as indicated in the experiment will be sent zeros.