Filtering implemented in ExperimentPrototype
Filtering info no longer coming from config. If not provided, a default
is created in the experiment. Modified sigprocpacket to include filter
and decimate data, as well as rxrate and output_sample_rate. Modified
datawritemetadata to have output_sample_rate. Filled these fields in
radar_control. The filtering information is set up according to classes
in the filtering module under experiment_prototype. The
ExperimentPrototype creates the default DecimationScheme if none is
provided.