write_bfiqdat added; noise_at_freq and ...
filtered_3db_bandwidth fields for conversion to iqdat, but also will be
useful for engineering purposes generally. Would like to try this method
of having separate bfiqdat and bfiq files because I don't believe that
all the acf information should really be placed in a basic bfiq file.
That file should solely be about the data (user can determine processing
methods). However, for compatability with current iqdat, we need a way
to write all fields so a separate bfiqdat write is necessary.