Fixed an error with comparing pointers to characters
In the first_one() function as well as the section of code where I was
comparing certain string elements to '1' in order to check agc and lp
status, I was comparing the pointer to '1'. Fixed this by first
declaring a new variables as characters, setting them to the value of
the element of pin_string, and using those for comparison. Should finish
build now.