Began rewriting timeline plotting function
Goal is for the plot to be a horizontally stacked bar plot, with each y
tick representing one day, with 24 hours on the x axis. This requires
the timeline events to be split if their duration takes them into the
next day. This will be done with the split_Event function, for which I
just finished an interface.