Changes to only output timeline list from convert_scd_to_timeline
Modified code to use a function that converts the timeline list to an
OrderedDict inplace when the ordered dictionary is needed. This is done
rather than outputting both the dictionary and the list. List is
converted to dictionary in plotting method for colour mapping, and in
the atq conversion method for scheduling.