Refactored writing of data and sequence timestamps
Rather than zero padding, I now find the maximum number of sequences of
any record in the file. This is used to create a buffer of all zeros
large enough to fit all the data if every record had the maximum number
of sequences. The data is then written to this buffer during the loop,
and reshaped at the end of execution. A similar procedure is used with
the sequence timestamps. This is done in both bfiq and pre-bfiq files.