Recursive function for removing problem antennas
Designed and implemented a recursive function to remove the antennas
with the most discrepancies associated with them. Do this until total
counted discrepancies is below a given total.

I think there's a better way to eliminate antennas and report on them.
This commit is for the sake of keeping track of history.