Adding some prelim work to fix this bug
I am attempting to see if keeping the first sequence synchronous will
remove this bug. I think that NVIDIA might have some hang ups when it
first loads sometimes that might cause a bottlenecked pipeline, so we
will see if this is enough to allow it to load correctly. The assertion
error doesnt seem to happen anytime other than startup.

Enabling more debug output for the time being.