Adding in code for intf offset
To be applied during beamforming. The centre offset along the array is
now accounted for in the interferometer beamforming (intf is beamformed
with the same phases as antennas 6,7,8,9 in main array in the Saskatoon
configuration). This did not work for all cases (sites) previously.

There was also a bug in how the centre was being calculated in the
get_phshift function, now fixed.