Modified iterators in radar_control
To properly account for INTTIME interfaced slices.

Some scans are aligned, where the start of the scan always
starts with the first beam, whereas others are not. Whether they are
aligned is dependent on if the total len(beam_order) = len(scanbound)
in the scan.

Scan class now has a aveperiod iterator for all possible aveperiods (sum
AvePeriod classes x beams within class), and aveperiods have their own
beam iterator. The beam iterator is only reset to 0 at the start of a scan
if the scan is aligned as above ^.

Removed check in experiment prototype of len(beam_order) = len(scanbound)
for each slice as this doesn't take into account INTTIME interfaced slices
and is not necessary if alignment isn't needed.