Logging (#377)
* introducing structlogs

* Created log_config.py
For each module we can import log_config and set a log level to begin logging

* log exceptions

* changing prints to log messages

* socket_operations.py logs
socket_operations.py now uses log.error() rather than taking a print shared macro.

* added stuctlogs
Updating to use structlogs, fixing comments spacings, and adding new log points.

* added graylog handler

* fixed scons build dir
scons was still using the head dir borealis/ when it should now be src/

* opt log handlers
Added options for log handlers in the config.ini ports. This allows us to enable/disable console, file, and aggregator forwarding logging. This will be handy if we decide printing to screen is not useful or streaming to the log aggregator uses too much bandwidth.

* remove copy config.ini
Made it so all the <site>_config.ini files can stay in config/ and do not need to be copied to the head dir when building. All options calls will use the RADAR_CODE environment variable and look in the correct directory. This was partially implemented before, but is now complete. Also added a dir for the lab at config/lab/.


---------

Co-authored-by: Lozzy-Bear <adam.lozinsky@usask.ca>